requirements.txt
allure-pytest
allure-python-commons~=2.13.5
pytest~=8.3.4
Faker~=37.3.0
jsonschema
httpx~=0.28.1
pydantic_core
pydantic~=2.11.4
pydantic-settings~=2.9.1
pydantic_core~=2.33.2
python-dotenv~=1.1.0
confluent-kafka~=2.10.0
lxml~=5.4.0
pydantic_core~=2.33.2